fix(tui): address unreviewed scroll/vim bugs from PR #1485 [Midtown !1752 !1753 !1754 !1755]
Four issues from riverside's review of #1485 that merged without being addressed:

- **!1752** Ctrl+D: add `!channel_switcher.show` guard matching Ctrl+U, so pressing
  Ctrl+D while the channel switcher overlay is open no longer scrolls behind it.

- **!1753** Vim j/k/g/G: add `focused_pane != InputBar` to the scroll gate so typing
  'got it', 'just a sec', or 'k sounds good' no longer loses the first character when
  InputBar is focused with an empty input buffer.

- **!1754** Test organization: move 10 inline scroll test functions out of the
  mod tests{} block into a new vim_scroll_tests.rs file, per CLAUDE.md convention.
  Replaces the test that codified the (now-fixed) wrong InputBar behavior.

- **!1755** Mouse scroll direction reversal: split `last_scroll_event` into separate
  `last_scroll_up_event` / `last_scroll_down_event` fields so a quick up->down reversal
  (within 50ms) enters immediate mode for the new direction rather than inertia mode.

All tests pass; 2 new direction-reversal tests added to mouse_scroll_tests.rs.
